hydra_elite_bot/requirements.txt
python-binance>=1.0.19
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
